ejercicio 9 tp3

+++ src/ar/edu/unq/poo/tp3/Punto.java
@@ -3,36 +3,39 @@ package ar.edu.unq.poo.tp3;
 import java.util.Objects;
 
 public class Punto {
-	
-	private float x;
-	private float y;
-	
+
+	private int x;
+	private int y;
+
 	public Punto() {
-		this.setXY(0,0);
+		this.setXY(0, 0);
 	}
-	
-	public Punto(float x, float y) {
+
+	public Punto(int x, int y) {
 		super();
-		this.setXY(x,y);
-		
+		this.setXY(x, y);
+
 	}
-	public void setXY(float x, float y) {
+
+	public void setXY(int x, int y) {
 		this.setX(x);
 		this.setY(y);
 	}
-	
-	public float getX() {
+
+	public int getX() {
 		return x;
 	}
-	public void setX(float x) {
+
+	public void setX(int x) {
 		this.x = x;
 	}
-	public float getY() {
+
+	public int getY() {
 		return y;
 	}
-	public void setY(float y) {
+
+	public void setY(int y) {
 		this.y = y;
 	}
-	
-	
+
 }
